fixed toast progress

+++ src/MobSwitcher.Core/Services/ITimerService.cs
@@ -6,6 +6,6 @@ namespace MobSwitcher.Core.Services
 {
     public interface ITimerService
     {
-        void Start(int minutes);
+        void Start(double minutes);
     }
 }
